# 🔒 Major Security Enhancement: Achieve 9.0/10 Security Score - Eliminate All Critical Vulnerabilities